+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master
 
+- Allow to `deny!` with details. ([@inkstak][])
+
 ## 0.7.1 (2024-07-25)
 
 - Support passing scope options to callable scope objects. ([@palkan][])